% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{make_kernel}
\alias{make_kernel}
\title{Generate a Polynomial Kernel}
\usage{
make_kernel(degree, R)
}
\arguments{
\item{degree}{the degree of the polynomial for the kernel}

\item{R}{the domain of the kernel}
}
\description{
Generate a Polynomial Kernel
}
