% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_blipCDF.R
\docType{data}
\name{Param_blipCDF}
\alias{Param_blipCDF}
\title{Blip CDF}
\format{\code{\link{R6Class}} object.}
\usage{
Param_blipCDF
}
\value{
\code{Param_base} object
}
\description{
Blip CDF
}
\section{Constructor}{

  \code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
    }
    \item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
    }
    \item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
    }
    \item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
    }
    \item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
    }
}
}

\concept{Parameters}
\keyword{data}
